+++ src/ModularPipelines.Docker/Options/DockerBuildOptions.cs
@@ -69,7 +69,7 @@ public record DockerBuildOptions([property: PositionalArgument(Position = Positi
     public string? AddHost { get; set; }
 
     [CommandSwitch("--build-arg")]
-    public KeyValueVariables? BuildArgs { get; set; }
+    public IEnumerable<KeyValue>? BuildArgs { get; set; }
 
     [CommandSwitch("--cache-from")]
     public string? CacheFrom { get; set; }