+++ src/ModularPipelines.Terraform/Options/TerraformImportOptions.cs
@@ -32,7 +32,7 @@ public record TerraformImportOptions([property: PositionalArgument(Position = Po
     public string? Provider { get; set; }
 
     [CommandSwitch("-var")]
-    public KeyValueVariables? Vars { get; set; }
+    public IEnumerable<KeyValue>? Vars { get; set; }
 
     [CommandSwitch("-var-file")]
     public string? VarFile { get; set; }